ENH: Allow for join between two multi-index dataframe instances (#20356)
