+++ pandas/tests/reshape/merge/test_merge.py
@@ -8,15 +8,14 @@ from datetime import date, datetime
 import numpy as np
 import pytest
 from numpy import nan
-from numpy.random import randn
 
 import pandas as pd
 import pandas.util.testing as tm
 from pandas import (Categorical, CategoricalIndex, DataFrame, DatetimeIndex,
-                    Float64Index, Index, Int64Index, MultiIndex, RangeIndex,
+                    Float64Index, Int64Index, MultiIndex, RangeIndex,
                     Series, UInt64Index)
 from pandas.api.types import CategoricalDtype as CDT
-from pandas.compat import lrange, lzip
+from pandas.compat import lrange
 from pandas.core.dtypes.common import is_categorical_dtype, is_object_dtype
 from pandas.core.dtypes.dtypes import CategoricalDtype
 from pandas.core.reshape.concat import concat
@@ -918,521 +917,6 @@ def _check_merge(x, y):
 
         # TODO check_names on merge?
         assert_frame_equal(result, expected, check_names=False)
-
-
-class TestMergeMulti(object):
-
-    def setup_method(self, method):
-        self.index = MultiIndex(levels=[['foo', 'bar', 'baz', 'qux'],
-                                        ['one', 'two', 'three']],
-                                labels=[[0, 0, 0, 1, 1, 2, 2, 3, 3, 3],
-                                        [0, 1, 2, 0, 1, 1, 2, 0, 1, 2]],
-                                names=['first', 'second'])
-        self.to_join = DataFrame(np.random.randn(10, 3), index=self.index,
-                                 columns=['j_one', 'j_two', 'j_three'])
-
-        # a little relevant example with NAs
-        key1 = ['bar', 'bar', 'bar', 'foo', 'foo', 'baz', 'baz', 'qux',
-                'qux', 'snap']
-        key2 = ['two', 'one', 'three', 'one', 'two', 'one', 'two', 'two',
-                'three', 'one']
-
-        data = np.random.randn(len(key1))
-        self.data = DataFrame({'key1': key1, 'key2': key2,
-                               'data': data})
-
-    def test_merge_on_multikey(self):
-        joined = self.data.join(self.to_join, on=['key1', 'key2'])
-
-        join_key = Index(lzip(self.data['key1'], self.data['key2']))
-        indexer = self.to_join.index.get_indexer(join_key)
-        ex_values = self.to_join.values.take(indexer, axis=0)
-        ex_values[indexer == -1] = np.nan
-        expected = self.data.join(DataFrame(ex_values,
-                                            columns=self.to_join.columns))
-
-        # TODO: columns aren't in the same order yet
-        assert_frame_equal(joined, expected.loc[:, joined.columns])
-
-        left = self.data.join(self.to_join, on=['key1', 'key2'], sort=True)
-        right = expected.loc[:, joined.columns].sort_values(['key1', 'key2'],
-                                                            kind='mergesort')
-        assert_frame_equal(left, right)
-
-    def test_left_join_multi_index(self):
-        icols = ['1st', '2nd', '3rd']
-
-        def bind_cols(df):
-            iord = lambda a: 0 if a != a else ord(a)
-            f = lambda ts: ts.map(iord) - ord('a')
-            return (f(df['1st']) + f(df['3rd']) * 1e2 +
-                    df['2nd'].fillna(0) * 1e4)
-
-        def run_asserts(left, right):
-            for sort in [False, True]:
-                res = left.join(right, on=icols, how='left', sort=sort)
-
-                assert len(left) < len(res) + 1
-                assert not res['4th'].isna().any()
-                assert not res['5th'].isna().any()
-
-                tm.assert_series_equal(
-                    res['4th'], - res['5th'], check_names=False)
-                result = bind_cols(res.iloc[:, :-2])
-                tm.assert_series_equal(res['4th'], result, check_names=False)
-                assert result.name is None
-
-                if sort:
-                    tm.assert_frame_equal(
-                        res, res.sort_values(icols, kind='mergesort'))
-
-                out = merge(left, right.reset_index(), on=icols,
-                            sort=sort, how='left')
-
-                res.index = np.arange(len(res))
-                tm.assert_frame_equal(out, res)
-
-        lc = list(map(chr, np.arange(ord('a'), ord('z') + 1)))
-        left = DataFrame(np.random.choice(lc, (5000, 2)),
-                         columns=['1st', '3rd'])
-        left.insert(1, '2nd', np.random.randint(0, 1000, len(left)))
-
-        i = np.random.permutation(len(left))
-        right = left.iloc[i].copy()
-
-        left['4th'] = bind_cols(left)
-        right['5th'] = - bind_cols(right)
-        right.set_index(icols, inplace=True)
-
-        run_asserts(left, right)
-
-        # inject some nulls
-        left.loc[1::23, '1st'] = np.nan
-        left.loc[2::37, '2nd'] = np.nan
-        left.loc[3::43, '3rd'] = np.nan
-        left['4th'] = bind_cols(left)
-
-        i = np.random.permutation(len(left))
-        right = left.iloc[i, :-1]
-        right['5th'] = - bind_cols(right)
-        right.set_index(icols, inplace=True)
-
-        run_asserts(left, right)
-
-    def test_merge_right_vs_left(self):
-        # compare left vs right merge with multikey
-        for sort in [False, True]:
-            merged1 = self.data.merge(self.to_join, left_on=['key1', 'key2'],
-                                      right_index=True, how='left', sort=sort)
-
-            merged2 = self.to_join.merge(self.data, right_on=['key1', 'key2'],
-                                         left_index=True, how='right',
-                                         sort=sort)
-
-            merged2 = merged2.loc[:, merged1.columns]
-            assert_frame_equal(merged1, merged2)
-
-    def test_compress_group_combinations(self):
-
-        # ~ 40000000 possible unique groups
-        key1 = tm.rands_array(10, 10000)
-        key1 = np.tile(key1, 2)
-        key2 = key1[::-1]
-
-        df = DataFrame({'key1': key1, 'key2': key2,
-                        'value1': np.random.randn(20000)})
-
-        df2 = DataFrame({'key1': key1[::2], 'key2': key2[::2],
-                         'value2': np.random.randn(10000)})
-
-        # just to hit the label compression code path
-        merge(df, df2, how='outer')
-
-    def test_left_join_index_preserve_order(self):
-
-        left = DataFrame({'k1': [0, 1, 2] * 8,
-                          'k2': ['foo', 'bar'] * 12,
-                          'v': np.array(np.arange(24), dtype=np.int64)})
-
-        index = MultiIndex.from_tuples([(2, 'bar'), (1, 'foo')])
-        right = DataFrame({'v2': [5, 7]}, index=index)
-
-        result = left.join(right, on=['k1', 'k2'])
-
-        expected = left.copy()
-        expected['v2'] = np.nan
-        expected.loc[(expected.k1 == 2) & (expected.k2 == 'bar'), 'v2'] = 5
-        expected.loc[(expected.k1 == 1) & (expected.k2 == 'foo'), 'v2'] = 7
-
-        tm.assert_frame_equal(result, expected)
-        tm.assert_frame_equal(
-            result.sort_values(['k1', 'k2'], kind='mergesort'),
-            left.join(right, on=['k1', 'k2'], sort=True))
-
-        # test join with multi dtypes blocks
-        left = DataFrame({'k1': [0, 1, 2] * 8,
-                          'k2': ['foo', 'bar'] * 12,
-                          'k3': np.array([0, 1, 2] * 8, dtype=np.float32),
-                          'v': np.array(np.arange(24), dtype=np.int32)})
-
-        index = MultiIndex.from_tuples([(2, 'bar'), (1, 'foo')])
-        right = DataFrame({'v2': [5, 7]}, index=index)
-
-        result = left.join(right, on=['k1', 'k2'])
-
-        expected = left.copy()
-        expected['v2'] = np.nan
-        expected.loc[(expected.k1 == 2) & (expected.k2 == 'bar'), 'v2'] = 5
-        expected.loc[(expected.k1 == 1) & (expected.k2 == 'foo'), 'v2'] = 7
-
-        tm.assert_frame_equal(result, expected)
-        tm.assert_frame_equal(
-            result.sort_values(['k1', 'k2'], kind='mergesort'),
-            left.join(right, on=['k1', 'k2'], sort=True))
-
-        # do a right join for an extra test
-        joined = merge(right, left, left_index=True,
-                       right_on=['k1', 'k2'], how='right')
-        tm.assert_frame_equal(joined.loc[:, expected.columns], expected)
-
-    def test_left_join_index_multi_match_multiindex(self):
-        left = DataFrame([
-            ['X', 'Y', 'C', 'a'],
-            ['W', 'Y', 'C', 'e'],
-            ['V', 'Q', 'A', 'h'],
-            ['V', 'R', 'D', 'i'],
-            ['X', 'Y', 'D', 'b'],
-            ['X', 'Y', 'A', 'c'],
-            ['W', 'Q', 'B', 'f'],
-            ['W', 'R', 'C', 'g'],
-            ['V', 'Y', 'C', 'j'],
-            ['X', 'Y', 'B', 'd']],
-            columns=['cola', 'colb', 'colc', 'tag'],
-            index=[3, 2, 0, 1, 7, 6, 4, 5, 9, 8])
-
-        right = DataFrame([
-            ['W', 'R', 'C', 0],
-            ['W', 'Q', 'B', 3],
-            ['W', 'Q', 'B', 8],
-            ['X', 'Y', 'A', 1],
-            ['X', 'Y', 'A', 4],
-            ['X', 'Y', 'B', 5],
-            ['X', 'Y', 'C', 6],
-            ['X', 'Y', 'C', 9],
-            ['X', 'Q', 'C', -6],
-            ['X', 'R', 'C', -9],
-            ['V', 'Y', 'C', 7],
-            ['V', 'R', 'D', 2],
-            ['V', 'R', 'D', -1],
-            ['V', 'Q', 'A', -3]],
-            columns=['col1', 'col2', 'col3', 'val'])
-
-        right.set_index(['col1', 'col2', 'col3'], inplace=True)
-        result = left.join(right, on=['cola', 'colb', 'colc'], how='left')
-
-        expected = DataFrame([
-            ['X', 'Y', 'C', 'a', 6],
-            ['X', 'Y', 'C', 'a', 9],
-            ['W', 'Y', 'C', 'e', nan],
-            ['V', 'Q', 'A', 'h', -3],
-            ['V', 'R', 'D', 'i', 2],
-            ['V', 'R', 'D', 'i', -1],
-            ['X', 'Y', 'D', 'b', nan],
-            ['X', 'Y', 'A', 'c', 1],
-            ['X', 'Y', 'A', 'c', 4],
-            ['W', 'Q', 'B', 'f', 3],
-            ['W', 'Q', 'B', 'f', 8],
-            ['W', 'R', 'C', 'g', 0],
-            ['V', 'Y', 'C', 'j', 7],
-            ['X', 'Y', 'B', 'd', 5]],
-            columns=['cola', 'colb', 'colc', 'tag', 'val'],
-            index=[3, 3, 2, 0, 1, 1, 7, 6, 6, 4, 4, 5, 9, 8])
-
-        tm.assert_frame_equal(result, expected)
-
-        result = left.join(right, on=['cola', 'colb', 'colc'],
-                           how='left', sort=True)
-
-        tm.assert_frame_equal(
-            result,
-            expected.sort_values(['cola', 'colb', 'colc'], kind='mergesort'))
-
-        # GH7331 - maintain left frame order in left merge
-        right.reset_index(inplace=True)
-        right.columns = left.columns[:3].tolist() + right.columns[-1:].tolist()
-        result = merge(left, right, how='left', on=left.columns[:-1].tolist())
-        expected.index = np.arange(len(expected))
-        tm.assert_frame_equal(result, expected)
-
-    def test_left_join_index_multi_match(self):
-        left = DataFrame([
-            ['c', 0],
-            ['b', 1],
-            ['a', 2],
-            ['b', 3]],
-            columns=['tag', 'val'],
-            index=[2, 0, 1, 3])
-
-        right = DataFrame([
-            ['a', 'v'],
-            ['c', 'w'],
-            ['c', 'x'],
-            ['d', 'y'],
-            ['a', 'z'],
-            ['c', 'r'],
-            ['e', 'q'],
-            ['c', 's']],
-            columns=['tag', 'char'])
-
-        right.set_index('tag', inplace=True)
-        result = left.join(right, on='tag', how='left')
-
-        expected = DataFrame([
-            ['c', 0, 'w'],
-            ['c', 0, 'x'],
-            ['c', 0, 'r'],
-            ['c', 0, 's'],
-            ['b', 1, nan],
-            ['a', 2, 'v'],
-            ['a', 2, 'z'],
-            ['b', 3, nan]],
-            columns=['tag', 'val', 'char'],
-            index=[2, 2, 2, 2, 0, 1, 1, 3])
-
-        tm.assert_frame_equal(result, expected)
-
-        result = left.join(right, on='tag', how='left', sort=True)
-        tm.assert_frame_equal(
-            result, expected.sort_values('tag', kind='mergesort'))
-
-        # GH7331 - maintain left frame order in left merge
-        result = merge(left, right.reset_index(), how='left', on='tag')
-        expected.index = np.arange(len(expected))
-        tm.assert_frame_equal(result, expected)
-
-    def test_left_merge_na_buglet(self):
-        left = DataFrame({'id': list('abcde'), 'v1': randn(5),
-                          'v2': randn(5), 'dummy': list('abcde'),
-                          'v3': randn(5)},
-                         columns=['id', 'v1', 'v2', 'dummy', 'v3'])
-        right = DataFrame({'id': ['a', 'b', np.nan, np.nan, np.nan],
-                           'sv3': [1.234, 5.678, np.nan, np.nan, np.nan]})
-
-        merged = merge(left, right, on='id', how='left')
-
-        rdf = right.drop(['id'], axis=1)
-        expected = left.join(rdf)
-        tm.assert_frame_equal(merged, expected)
-
-    def test_merge_na_keys(self):
-        data = [[1950, "A", 1.5],
-                [1950, "B", 1.5],
-                [1955, "B", 1.5],
-                [1960, "B", np.nan],
-                [1970, "B", 4.],
-                [1950, "C", 4.],
-                [1960, "C", np.nan],
-                [1965, "C", 3.],
-                [1970, "C", 4.]]
-
-        frame = DataFrame(data, columns=["year", "panel", "data"])
-
-        other_data = [[1960, 'A', np.nan],
-                      [1970, 'A', np.nan],
-                      [1955, 'A', np.nan],
-                      [1965, 'A', np.nan],
-                      [1965, 'B', np.nan],
-                      [1955, 'C', np.nan]]
-        other = DataFrame(other_data, columns=['year', 'panel', 'data'])
-
-        result = frame.merge(other, how='outer')
-
-        expected = frame.fillna(-999).merge(other.fillna(-999), how='outer')
-        expected = expected.replace(-999, np.nan)
-
-        tm.assert_frame_equal(result, expected)
-
-    def test_join_multi_levels(self):
-
-        # GH 3662
-        # merge multi-levels
-        household = (
-            DataFrame(
-                dict(household_id=[1, 2, 3],
-                     male=[0, 1, 0],
-                     wealth=[196087.3, 316478.7, 294750]),
-                columns=['household_id', 'male', 'wealth'])
-            .set_index('household_id'))
-        portfolio = (
-            DataFrame(
-                dict(household_id=[1, 2, 2, 3, 3, 3, 4],
-                     asset_id=["nl0000301109", "nl0000289783", "gb00b03mlx29",
-                               "gb00b03mlx29", "lu0197800237", "nl0000289965",
-                               np.nan],
-                     name=["ABN Amro", "Robeco", "Royal Dutch Shell",
-                           "Royal Dutch Shell",
-                           "AAB Eastern Europe Equity Fund",
-                           "Postbank BioTech Fonds", np.nan],
-                     share=[1.0, 0.4, 0.6, 0.15, 0.6, 0.25, 1.0]),
-                columns=['household_id', 'asset_id', 'name', 'share'])
-            .set_index(['household_id', 'asset_id']))
-        result = household.join(portfolio, how='inner')
-        expected = (
-            DataFrame(
-                dict(male=[0, 1, 1, 0, 0, 0],
-                     wealth=[196087.3, 316478.7, 316478.7,
-                             294750.0, 294750.0, 294750.0],
-                     name=['ABN Amro', 'Robeco', 'Royal Dutch Shell',
-                           'Royal Dutch Shell',
-                           'AAB Eastern Europe Equity Fund',
-                           'Postbank BioTech Fonds'],
-                     share=[1.00, 0.40, 0.60, 0.15, 0.60, 0.25],
-                     household_id=[1, 2, 2, 3, 3, 3],
-                     asset_id=['nl0000301109', 'nl0000289783', 'gb00b03mlx29',
-                               'gb00b03mlx29', 'lu0197800237',
-                               'nl0000289965']))
-            .set_index(['household_id', 'asset_id'])
-            .reindex(columns=['male', 'wealth', 'name', 'share']))
-        assert_frame_equal(result, expected)
-
-        assert_frame_equal(result, expected)
-
-        # equivalency
-        result2 = (merge(household.reset_index(), portfolio.reset_index(),
-                         on=['household_id'], how='inner')
-                   .set_index(['household_id', 'asset_id']))
-        assert_frame_equal(result2, expected)
-
-        result = household.join(portfolio, how='outer')
-        expected = (concat([
-            expected,
-            (DataFrame(
-                dict(share=[1.00]),
-                index=MultiIndex.from_tuples(
-                    [(4, np.nan)],
-                    names=['household_id', 'asset_id'])))
-        ], axis=0, sort=True).reindex(columns=expected.columns))
-        assert_frame_equal(result, expected)
-
-        # invalid cases
-        household.index.name = 'foo'
-
-        def f():
-            household.join(portfolio, how='inner')
-
-        pytest.raises(ValueError, f)
-
-        portfolio2 = portfolio.copy()
-        portfolio2.index.set_names(['household_id', 'foo'])
-
-        def f():
-            portfolio2.join(portfolio, how='inner')
-
-        pytest.raises(ValueError, f)
-
-    def test_join_multi_levels2(self):
-
-        # some more advanced merges
-        # GH6360
-        household = (
-            DataFrame(
-                dict(household_id=[1, 2, 2, 3, 3, 3, 4],
-                     asset_id=["nl0000301109", "nl0000301109", "gb00b03mlx29",
-                               "gb00b03mlx29", "lu0197800237", "nl0000289965",
-                               np.nan],
-                     share=[1.0, 0.4, 0.6, 0.15, 0.6, 0.25, 1.0]),
-                columns=['household_id', 'asset_id', 'share'])
-            .set_index(['household_id', 'asset_id']))
-
-        log_return = DataFrame(dict(
-            asset_id=["gb00b03mlx29", "gb00b03mlx29",
-                      "gb00b03mlx29", "lu0197800237", "lu0197800237"],
-            t=[233, 234, 235, 180, 181],
-            log_return=[.09604978, -.06524096, .03532373, .03025441, .036997]
-        )).set_index(["asset_id", "t"])
-
-        expected = (
-            DataFrame(dict(
-                household_id=[2, 2, 2, 3, 3, 3, 3, 3],
-                asset_id=["gb00b03mlx29", "gb00b03mlx29",
-                          "gb00b03mlx29", "gb00b03mlx29",
-                          "gb00b03mlx29", "gb00b03mlx29",
-                          "lu0197800237", "lu0197800237"],
-                t=[233, 234, 235, 233, 234, 235, 180, 181],
-                share=[0.6, 0.6, 0.6, 0.15, 0.15, 0.15, 0.6, 0.6],
-                log_return=[.09604978, -.06524096, .03532373,
-                            .09604978, -.06524096, .03532373,
-                            .03025441, .036997]
-            ))
-            .set_index(["household_id", "asset_id", "t"])
-            .reindex(columns=['share', 'log_return']))
-
-        def f():
-            household.join(log_return, how='inner')
-
-        pytest.raises(NotImplementedError, f)
-
-        # this is the equivalency
-        result = (merge(household.reset_index(), log_return.reset_index(),
-                        on=['asset_id'], how='inner')
-                  .set_index(['household_id', 'asset_id', 't']))
-        assert_frame_equal(result, expected)
-
-        expected = (
-            DataFrame(dict(
-                household_id=[1, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 4],
-                asset_id=["nl0000301109", "nl0000289783", "gb00b03mlx29",
-                          "gb00b03mlx29", "gb00b03mlx29",
-                          "gb00b03mlx29", "gb00b03mlx29", "gb00b03mlx29",
-                          "lu0197800237", "lu0197800237",
-                          "nl0000289965", None],
-                t=[None, None, 233, 234, 235, 233, 234,
-                   235, 180, 181, None, None],
-                share=[1.0, 0.4, 0.6, 0.6, 0.6, 0.15,
-                       0.15, 0.15, 0.6, 0.6, 0.25, 1.0],
-                log_return=[None, None, .09604978, -.06524096, .03532373,
-                            .09604978, -.06524096, .03532373,
-                            .03025441, .036997, None, None]
-            ))
-            .set_index(["household_id", "asset_id", "t"]))
-
-        def f():
-            household.join(log_return, how='outer')
-
-        pytest.raises(NotImplementedError, f)
-
-    @pytest.mark.parametrize("klass", [None, np.asarray, Series, Index])
-    def test_merge_datetime_index(self, klass):
-        # see gh-19038
-        df = DataFrame([1, 2, 3],
-                       ["2016-01-01", "2017-01-01", "2018-01-01"],
-                       columns=["a"])
-        df.index = pd.to_datetime(df.index)
-        on_vector = df.index.year
-
-        if klass is not None:
-            on_vector = klass(on_vector)
-
-        expected = DataFrame(
-            OrderedDict([
-                ("a", [1, 2, 3]),
-                ("key_1", [2016, 2017, 2018]),
-            ])
-        )
-
-        result = df.merge(df, on=["a", on_vector], how="inner")
-        tm.assert_frame_equal(result, expected)
-
-        expected = DataFrame(
-            OrderedDict([
-                ("key_0", [2016, 2017, 2018]),
-                ("a_x", [1, 2, 3]),
-                ("a_y", [1, 2, 3]),
-            ])
-        )
-
-        result = df.merge(df, on=[df.index.year], how="inner")
-        tm.assert_frame_equal(result, expected)
 
 
 class TestMergeDtypes(object):
